+++ src/application/editor/EditorGroup.jsx
@@ -7,7 +7,7 @@ import "../css/Editor.css";
 import Textarea from "../ui/Textarea";
 import Renderer from "./Renderer";
 import EditorGroupBar from "./EditorGroupBar";
-import FileSystem from "../explorer/FileSystem";
+import FileSystem from "../FileSystem";
 
 const style = {
     container: {
