+++ src/application/settings/UserPreferences.jsx
@@ -1,5 +1,5 @@
 import { toast } from "react-toastify";
-import FileSystem from "../explorer/FileSystem";
+import FileSystem from "../FileSystem";
 // import { toast } from 'react-toastify';
 
 export default class UserPreferences {
@@ -58,12 +58,6 @@ export default class UserPreferences {
     }
 
     static set(key, value) {
-        if (key === "userStorage" && typeof value !== "string") {
-            toast.error("Tryin to store true again?!", {
-                autoClose: false,
-            });
-            return;
-        }
         let contents = this.getPreferences();
         if (contents.error === true) return;
         contents[key] = value || true;
